requirements.txt
appnope==0.1.3
asgiref==3.3.1
asttokens==2.4.0
backcall==0.2.0
black==23.10.0
click==8.1.7
configparser==6.0.0
decorator==5.1.1
Django==3.1.8
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
executing==2.0.0
ipython==8.13
isort==5.12.0
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
stack-data==0.6.3
traitlets==5.11.2
uWSGI==2.0.22
wcwidth==0.2.8
